Refactor Module and Morphism into separate modules
