+++ quiver_representations/__init__.py
@@ -1,5 +1,6 @@
 from .quiver import Quiver
 from .field import Field, FiniteField, ComplexNumbers, ZeroMap
-from .representation import Module, Morphism
+from .module import Module
+from .morphism import Morphism
 
 __all__ = ['Quiver', 'Field', 'FiniteField', 'ComplexNumbers', 'Module', 'ZeroMap', 'Morphism']
